Usability fixes encountered during testing
- Report a useful error message for docker if we're missing an image and
  don't have a way to retrieve it.
- When running locally, pass environmental variables along to the
  subprocess run. This keeps local PATHs and friends working.
- Avoid removing temporary directories if '--leave-tmpdir' is set.

+++ reference/cwltool/workflow.py
@@ -284,8 +284,9 @@ class Workflow(Process):
 
             for a in actual_jobs:
                 if a.outdir:
-                    _logger.info("Removing intermediate output directory %s", a.outdir)
-                    shutil.rmtree(a.outdir, True)
+                    if kwargs.get("rm_tmpdir", True):
+                        _logger.info("Removing intermediate output directory %s", a.outdir)
+                        shutil.rmtree(a.outdir, True)
 
         output_callback(wo, self.processStatus)
